refactor(agent): split capability modules
