+++ packages/agent/src/capabilities/workspace-shell.ts
@@ -1,0 +1,22 @@
+import {
+  defineCapability,
+  normalizeMode,
+} from "../capability-runtime.ts"
+
+import type {
+  AgentCapabilityDefinition,
+  AgentCapabilityMode,
+  AgentToolSet,
+} from "../types.ts"
+
+export function workspaceShell(options: { mode?: AgentCapabilityMode } = {}): AgentCapabilityDefinition {
+  const mode = normalizeMode(options.mode, "Workspace Shell")
+  return defineCapability({
+    id: "workspace-shell",
+    mode,
+    requires: [{ primitive: "workspace", workspace: { mode, required: true } }],
+    tools: ({ workspace }) => (mode === "write" && "write" in workspace.tools
+      ? (workspace.tools as unknown as { write: () => AgentToolSet }).write()
+      : workspace.tools.inspect()) as AgentToolSet,
+  })
+}
